Create ethereum.py
Imports the necessary libraries: Web3 for connecting to an Ethereum node, HTTPProvider for connecting to a remote node via HTTP, compile_source from solcx for compiling the smart contract source code, and json for working with JSON data.

Connects to an Ethereum node using the HTTPProvider class and an Infura project ID. You'll need to replace your_project_id with your actual Infura project ID.

Defines a contract_source string that contains the source code for a simple Solidity smart contract that stores and retrieves a message.

Compiles the smart contract source code using the compile_source function from solcx. This produces a dictionary of compiled contracts.

Extracts the ABI (Application Binary Interface) and bytecode from the compiled contract dictionary.

Deploys the smart contract to the Ethereum blockchain using the constructor method of the Contract class. This sets the initial message value to "Hello, World!".

Waits for the transaction to be mined and returns the transaction receipt. This contains information about the newly deployed contract, including its address on the Ethereum blockchain.

Interacts with the smart contract by creating an instance of the Contract class using the contract address and ABI.

Retrieves the current message value using the getMessage function of the smart contract.

Updates the message value using the setMessage function of the smart contract.

Retrieves the updated message value using the getMessage function of the smart contract.

Prints the current and updated message values to the console.