what it does
This script defines a set of utility functions that can be used throughout the CLI tool. Here's a brief explanation of each function:

validate_address(): Validates the format of an Ethereum address. If the address is invalid, a click.BadParameter exception is raised.

validate_private_key(): Validates the format of an Ethereum private key. If the private key is invalid, a click.BadParameter exception is raised.

validate_contract_address(): Validates the format of an Ethereum contract address. If the contract address is invalid, a click.BadParameter exception is raised.

get_gas_price(): Gets the current gas price for the specified network.

get_gas_limit(): Gets the default gas limit for contract transactions.

get_account(): Gets an account object from the specified private key.

get_contract_interface(): Gets the ABI and bytecode for the specified contract.

load_contract_instance(): Loads a contract instance with the specified ABI and address.